requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
brotlipy==0.7.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==22.9.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1649105784853/work
contourpy==1.0.5
cryptography @ file:///tmp/build/80754af9/cryptography_1639414572950/work
cycler==0.11.0
docker-pycreds==0.4.0
fire==0.4.0
fonttools==4.37.4
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.9
GitPython==3.1.29
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.22.3
iteration-utilities==0.11.0
joblib==1.2.0
kiwisolver==1.4.4
lightning-utilities==0.5.0
llvmlite==0.39.1
matplotlib==3.6.1
medmnist==2.1.0
multidict==6.0.4
networkx==2.8.8
numba==0.56.4
numpy==1.23.4
packaging==21.3
pandas==1.5.0
pathtools==0.1.2
Pillow==9.2.0
promise==2.3
protobuf==4.21.8
psutil==5.9.3
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pynndescent==0.5.8
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytorch-lightning==1.9.0
pytz==2022.5
PyWavelets==1.4.1
PyYAML==6.0
qpsolvers==2.3.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016711199/work
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.2
sentry-sdk==1.9.10
setproctitle==1.3.2
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
smmap==5.0.0
termcolor==2.1.0
threadpoolctl==3.1.0
tifffile==2022.10.10
toolz @ file:///croot/toolz_1667464077321/work
torch==1.12.1
torch-summary==1.4.5
torchmetrics==0.11.0
torchvision==0.13.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
typing_extensions==4.4.0
umap-learn==0.5.3
urllib3 @ file:///croot/urllib3_1666298941550/work
wandb==0.13.4
yarl==1.8.2
